Modify encoder for use as a companion to part 1 of the blog series
In part 1 of the associated blog series, we discuss the structure
needed to encode an AVIF file, culminating in a simple gray-square
output file. Everything to do with encoding a specific image is
left for a future post.

To help readers match the discussion in that post to code, this
commit strips down the encoder to only the parts discussed in that
blog post.

Compared to the full version of tinyavif, this commit makes the
following changes:

* Rework the core encoder (TileEncoder) to generate all 64x64 blocks,
  all marked skip

* Remove prediction, transform, and reconstruction functions

* Remove coefficient encoding and associated CDFs

* Remove input file processing

* Modify the handful of CDFs which are left to fit the different
  subset of contexts that we end up encountering